Refurbished options and options retrieval
Options of container are now responsible of computing their related
types and values. Moreover, the options fetching code is simplified
and put back the semantic tag in the spotlight.

Semantic tag (like tag::table_) now defines the actual mapping process
and defaulting process of option fetching. This lead to the upcoming
change of itnerface in container and related systems.

Unit Tests on options have been streamlined and completed.